Adapt Tugboat branch test for dash shell
Tugboat uses `dash` which uses a single `=` for string comparison rather
than bash' double `==`. This is somewhat confusing since the default
login shell for our terminal user is actually bash and the log output is
`/bin/sh` which would usually be a shell that supports `==` but is
instead symlinked to dash.