Simplify test comment creation
The comment module is a bit weird, especially when dealing with
permissions and using the current user correctly is important. So we use
the current user as author rather than specifying the author directly to
ensure that we use the permissions of the user we expect.

This also revealed we were running into a Drupal core bug:
https://www.drupal.org/project/drupal/issues/3039955