Split up comment access test cases
This creates slightly more tests although these tests are cheap to run.
This makes the entire class slightly more verbose but the benefit is
that breakages should more clearly indicate what went wrong, rather than
having multiple options of mistakes.

This also uses the `CommentInterface::*` constants because our tests
don't actually care about the specific values and using the constants
makes the test clearer.

We also don't rely on the default value for the status because what
we're testing shouldn't change if the default value changes.