Improve robustness of SocialProfileFieldsOverride
The override accesses existing configuration which can fail during
installation when that configuration may not yet be available.
Additionally we've learned since writing this class that we can use the
string value as a unique key to ensure that deep merging works. This
ensures we don't need to count the existing configuration which slightly
speeds up the override (and makes it more robust as a whole).