Enable query access checks in tests without social_core
The improvements to the comment access layer require a setting from
social_core to be set. For tests that module requires too much and
brings along too much config to be easily installed, therefore we set
the setting manually.

The GraphQL module contains an install hook that sets the setting for us
(given social_core, it's dependency, is enabled in a normal site
install) on production sites.