Rename SocialComments to QueryComments
This makes the name of the data producer more closely fit its expected
place in the GraphQL schema (`Query.comments`). It also now followes
`UserComment` which does the same (albeit without the plurality it
should have). Finally it now also matches the `QueryCommentsTest` so
that its easy to know what's being tested.