Merge pull request #2241 from goalgorilla/issue/3198737-constraint-violation-multilanguage-group-content
Issue #3198737 by xvendo, ronaldtebrake: Ensure when group content is translated we don't update the group content entirely, this is redundant. We also ensure we do not allow for changing group references and visibility for now, as they can create unwanted behavior on editing group content with duplicate access records as it seems like a new group is added but that is not true.