Move PHPUnit testing from Travis to GitHub Actions
This removes PHPUnit testing from Travis and instead move it to GitHub
Actions.

PHPUnit requires a proper Drupal set-up to execute so we add a
composer.json project file for testing. The GitHub action downloads this
test file and then configures composer to install the latest version of
the code for which the workflow is being run.

GitHub Actions runs PHPUnit directly rather than through run-tests in
social_scripts. This decouples the configuration of Open Social's CI
from other repositories. Additionally since XDebug is not installed on
our GitHub actions container we don't need long PHP config arguments to
disable it.

The PHPUnit configuration is updated to use a SQLite database which
works for all types of tests except browsertests (which requires more
infrastructure). This avoids the need for an external MySQL database and
has the added benefit of speeding up tests because database requests
don't need to go over the network.