Return empty profile when no default profile could be loaded
Our profile field isn't allowed to return NULL because all users should
have a profile (and profiles aren't accessible as standalone data) so if
a user is not allowed to view profile data then we just fake this by
returning an empty profile instead. The profile fields are already
nullable because users can choose visibility on a per-field basis.