Refactor profile list and field access checks
This moves the access check for `mail` and `langcode` which we deemed
priviliged into field access checks which allows other Drupal systems to
more easily check them.

Additionally this introduces a new `list user` (in the form `list
<entity>`) permission ahead of
https://www.drupal.org/project/drupal/issues/2953566 that is required
for users to create a listing of users.

The permission to view users and profiles is needed for every user of
Open Social since we otherwise can't render things like actvitities for
anonymous or authenticated users. However, we have custmo access checks
on overview pages to prevent anonymous users from easily listing all
users on a platform. Such access check was not yet available as
permission which meant that the GraphQL API could inadvertently list all
users.