Add visibility fields to user profiles
This introduces logic to duplicate all fields on a profile with a
visibility field. This allows us to move our per-profile visibility
settings out of user.data and into the profile itself. This makes it
possible to query the visibility settings while fetching data to provide
simpler and more robust access control for profile fields.

As a first step the ProfileEntitySortable views field sorter is
rewritten to use this new structure. This illustrates how we can use the
new fields in a query directly rather than having to jump through hoops
with complex pre-calculated fields.

This commit also introduces the email and preferred_language profile
fields. This duplicates data from the user entity onto the profile over
which we also want visibility controls. By implementing an automatic
sync from user entity into the profile we can keep these settings
managed on the user entity but treat it like any other profile
information for view access and rendering.

The new profile field permissions introduced allow us to assign the
different per-field visibility levels to roles. For example by default
(not included in this commit) Site Managers should receive the ability
to see private fields but we may also want to create an HR role that
also has this permission but can not change platform settings.
Additionally some platforms may want community fields visible for any
authenticated user but others may only want it for verified users.

The friends visibility is introduced to allow a future setting between
"private" and "community". The query for this can be adjusted to load
the friends of the viewing user and allow visibility based on the
current conditions OR when the owner of the viewed profile is in the
viewer's list of friends.