Issue #3185523 by rolki, Kingdutch: Implement public Topic querying in GraphQL
Allows filtering by topic type and provides test coverage to ensure that
querying requires the correct permissions to view the content.

The URL data producer in GraphQL doesn't yet allow producing absolute
URLs. However, this is needed when making content available for external
platforms. A pull request is opened to the GraphQL module itself in
https://github.com/drupal-graphql/graphql/issues/1219 however adding
that as a patch may cause patch issues when a new module is released so
we include the DataProducer as deprecated in Open Social for now. This
allows the functionality to be used with a clear future upgrade path.

External modules should not rely on this data producer but instead use
the one from the GraphQL module if it's available or provide their own
in the meantime.