Create visibility field configs rather than a bundle info hook
In order to change the default value for a field it must have an actual
changeable config. This means that the earlier method of simply
implementing a hook_bundle_field_info didn't work anymore (because it's
static).

As a result the update hook must be altered and code is added to keep
the field configs in sync as well. We do this based on the creation of
field configs for the fields we're managing rather than doing it in
lockstep with the creation of field storage.