Migrate profile submodule data into social_profile
This commit contains all the update hooks that actually migrate the data
from the old modules into the new field and permission structure.

As final step the old submodules are uninstalled. To prevent other
modules from being installed accidentally we disallow starting the
update if any enabled modules still depend on a module that would be
disabled.