Re-implement default profile creation
Previously we only created the a profile if it was configured not to
create one during registration. However, in the latest changes we always
create a profile during registration. There are also scenarios that were
now unhandled in which a user was created through other means than user
registration in which case a profile was not created.

This commit ensures that a profile is always created regardless of
profile module configuration.