Fix non-namespaced class in use statement
As coding standards indicate we should not add `use` statements for
non-namespaced classes. However, referencing these directly will make
PHP think that we're looking for the class within the current namespace
of the file. The solution is to use a backslash (`\`) before the
class-name to indicate the class can be found in the empty namespace.