Use own copy of `GraphQLTestBase` in tests
The `GraphQLTestBase` of the graphql module creates a user during set-up
and installs languages. This may be problematic in case we have a module
that acts on this (e.g. through hook_user_insert) and then relies on
their own configuration that may not yet be installed. This happens
because all modules are installed before `setUp` is called (enabling
their hooks) but the GraphQL module's `setUp` is called before some of
our own config is installed (to ensure dependent config is installed in
the proper order).

The easiest way to solve this is to duplicate everything we need but
remove the offending lines. We already expected tests to be explicit
about setting up their test user in the class. We choose to keep a
separate class so that it's clear what is part of the `graphql` module's
dependencies and what is needed for Open Social's added functionality.
This should also help keep the class up-to-date during module updates.