Expand PHPStan coverage
The selected folders were excluded because these classes needed
PHP 7.4 and we were testing on PHP 7.3 which caused errors.

Now that Open Social uses PHP 7.4 as minimum version we need to
ensure that we can't commit code to those folders without
PHP standards being applied.

The phpstan PHPUnit extension is installed together with the extension
autoloader to make our configuration slightly easier.

PHPUnit is installed because PHPStan looks at composer's autoloader to
find classes and if PHPUnit is not a dev dependency then it won't show
up in the autoloader. This ensures we can enable converage for our test
classes too.