Move Travis tests to PHP 7.4
Updating the minimum version of PHP to 7.4 means that we also need to
update the version of PHP that we use in testing, otherwise composer will
not install.

This means that for backported features developers must make sure that
they do not use any PHP 7.4 only features on their own in their backports.

As part of this change we upgrade the Ubuntu version that we build on
since this is needed to get a PHP 7.4 image that includes the extensions
that Drupal needs.