Adjust user test to new GraphQL base class
The base class no longer creates an initial user because this was
causing issues during installation. As a result we must adjust tests
that depend on what users are stored in the database. We now have more
control but must thus also be more secure in what user we're viewing as.