Use `Url` over `string`
We should prefer using classes that concretely define types and
validations over strings that can hold arbitrary values. This ensures
that we don't end up with invalid URLs in our e-mail call to actions. It
also allows us to remove some checks to ensure we don't have an empty
string (since the typesystem now takes care of that).