Add script that can be run as pre-commit hook
The script makes it easy to execute all the things that we also run in
our GitHub workflows. This makes it easier to check those things
locally.

It can also easily be used in `git bisect` or `git rebase --exec` to
find a broken commit or verify that each commit is working.