Enable workflows on push to main
To allow external services such as Weblate or our translation extraction
to push to main we must also run our workflows for those events. If we
don't do so then the status will default to failed and the push will be
rejected.

While making this change we unify our workflows towards one for coding
standards and one for testing. Those can simply run multiple jobs for
the type that there is.

In unifying the coding standards we remove some php-actions we don't
need and standardise on selecting PHP 7.4 within the container we're
running in. This makes usage slightly more as a developer would do from
the command line and requires less documentation reading to understand.