Replace mglaman/drupal-check my PHPStan rules
mglaman/drupal-check provides a separate command to check for
deprecations in the codebase. We weren't actually using this command in
our CI or workflows. However, we were implicitly loading the PHPStan
library that it relied on in our PHPStan execution.

Replacing the tool with a requirement on the actual PHPStan ruleset
removes some unneeded packages and makes our coding standards more
explicit.