Provide test coverage for revamped profile functionality
This commit introduces test coverage for the new profile functionality
and the related access checks. While doing so some changes are made
to resolve issues that were uncovered by the tests.

The UserSchema mapping is changed from the `Path` DataProducer to
instead use the `Field` data producer. The latter properly calls the
access check method. This change requires us to compose the data
producer with the `Path` data producer to fetch the value after the
field is access checked.

An important change in tests is to ensure that `setUpCurrentUser` is
called at the start of each test dealing with users. The method ensures
that ‘user 1’ is created, even though it may not be created in tests.
Not calling `setUpCurrentUser` would cause the first user created with
`createUser` to have uid 1 and this can cause access checks to pass when
they shouldn’t.

The GraphQLUsersEndpointTest is updated to indicate that the
`social_user`  module itself does not protect the email. Instead it
defers to Drupal’s default behaviour. The `social_profile` adds
per-field permissions that also properly protects the email field.

To ensure that the API can access `mail` on the `User` entity the
`social_profile` module applies the access checks for the email profile
field also to the mail field on the User entity. The email is not moved
to the GraphQL API because we may want to differentiate between a
visible profile email and an account email in the future. Currently the
account email is also visible on the profile.

The field access checks in social_profile are abstracted to show that
they work on other Fieldable entity types too. There’s nothing specific
about the Profile entity. For now though it was decided not to move
this functionality into social_core because we could not come up with a
use-case outside of profiles. However, this should be an easy step in
the future.