Replace SocialUserHelper by User entity class method
We already overwrite the User class so we can make the check far easier
to consume that having to call an external helper.

Additionally this now explicitly checks for the `verified` role to know
if a user is verified. This more closely matches other parts of the
changes and allows adding roles to unverified users.