Require user view permission to view profiles
Before this commit we had quite a few places that would manually check
whether the user had access to view the user (e.g. because the user may
be blocked or unverified) before granting access to the profile.

This commit clearly writes down that decision by implementing an access
handler.

This also allows us to clean up the UserProfileController access check
somewhat. We can not use the `profile.view` entity access check because
at the time the access handler is run we only have a user and not a
profile. This can be fixed with a custom parameter upcaster but that's
more effort than it solves.