Add initial install permissions for profiles
This commit ensures that when Open Social is installed the permissions
around profile fields are properly set.

For fields that are added later we only set the site manager permissions
that the site manager can not change. This means that the update hook
that introduces the field must decide what permissions for normal users
should be set.