Remove social_user update hooks pre-10.0
This appears to change an update hook that was left behind as well but
it only moves the hook_update_dependencies implementation to the top.
There's no functional change to tests.

The old user permission function is removed since it was used in a test
(which is a bad idea) and should not have been called by other modules.