Fix visibility settings not editable on new install
A user on a fresh Open Social installation was not able to change their
field visibility settings. This was because the profile sub-form was
added in an update hook but was not included in any installable config
or install hook.

An install hook function is added. As far as I'm aware it's not possible
to add this field to the default user.user.default view mode that we
ship with Open Social (that would further couple the installation
profile to social_profile).