Move overwritten admin/people view to social_profile
The admin view had a hidden dependency on social_profile. This was
originally added in 9a1dc5e1. The problem is that for other installer
related issues we now actually declare the dependency from
`social_profile` on `social_user`. This causes the first name and last
name fields not to exist when `social_user` is installed.

To solve this we move the view to the social_profile module. The default
Drupal view works fine if we only have user entities. Only once we add
custom profile information do we want to update this view, this
coincides with installing `social_profile`.