Added patch on redirect module for allowing redirection when the interface language is different from content language.
On a multilingual website, when a user creates a content in one language and then moves it to a group, a redirect is registered for that path. But when the old link is circulated, and shared with user from different language the content is not found as the redirect is not registered for that language.