Fix old module uninstalling for profile functionality
If you give Drupal a bunch of modules to uninstall it'll only do so if
they are all installed. This means we actually need to check whether
they're present because we want to end up with none of them installed
but they're optional, so it's possible that only one (or none) is
actually enabled.