Issue #3264287 by robertragas, Ressinel: Use the hook to remove users from notifications.
So we use the new private_message module version which turns the private message notifier into a submodule, but in the submodule, they actually change the method to contain 2 arguments instead of 3, and the one they removed $members is the one we actually alter. This has been moved to an alter hook, so we don't actually need the service anymore and move it to the alter hook.