Issue #3263114 by Ressinel: Make sure that users can't accept invites if there are no spots left.
Issue #3263114 by Ressinel: Event creator should not be able to set 'Maximum enrollment' to 0, it doesn't make sense, and 1 should be the default value.

Issue #3263114 by Ressinel: Make sure that managers not able to add enrolments if there are no spots left.

This change is necessary because this particular class is not extending any class that guarantees that "routeMatch" attribute will be an object.