Adding suggested code review changes
Adding missing permission to hook_install from social_core and a new line at the end of the svg file. Change the menu icon to use the cog svg and changed the CSS code to match the current theme color. Added the option to better select the homepage block, based at the criteria that homepage blocks are set to anonymous users, and placed at the current theme hero region.