Re-remove SocialProfileNameService
This was previously removed as part of the work on the profile refactor
but a bug was "fixed" in EnrollRequestDeclineForm which accidentally
caused the class to be re-introduced in the rebase.

Rather than using this special service, we should just call
`getDisplayName` on the account, which was build to do exactly what is
desired.

The PR where the EnrollRequestDeclineForm was bothced is in
https://github.com/goalgorilla/open_social/pull/2794

The form itself needs some more love because it doesn't actually do much
without an enroll request, but we're already fixing too much in one
branch so I'll leave that to someone else.