Issue #3273900 by Ressinel: Update socialblue dependency to 2.1.4
Until 2.1.3 socialblue had a non-fixed dependency for socialbase base theme.
This could cause issues where users were able to download a socialbase version that was not
intended for the release version of socialblue and Open Social.

The latest socialblue release fixes the socialbase theme version, and this is why we are
updating socialbase dependency on Open Social

See: https://www.drupal.org/project/socialblue/issues/3273685