Update composer.json and tests/composer.json
This change is required because now we need to define custom packages and add them to our composer.json repositories key. This is crucial due to the removal of asset-packagist.

Please note the following for our composer.json files:

open_social/composer.json used for PHPSTAN and PHPCS.
open_social/tests/composer.json used for Tugboat.