Issue #3275980 by navneet0693: Switching language on multilingual platform leads to unexpected error
Refactored the code to accept few more scenarios when limiting the current accounts to perform CRUD operations.

For example, we don't this code to block any one with 'bypass social_core menu links access' permission. This permission indicates that user is a super user.

Also, this code is targetted towards CM/SM users which have 'administer menu' so a check was also added to take that into consideration. The code refactored to take postive case scenarios into to consideration instead of negative because there is more chance of traffic from such users without the above mentioned permissions, which also includes anonymous users traffic.

Adding the drupal_static functionality to _social_core_default_main_menu_links() helps to make the code more efficient as the scenario to change it really low.