Issue #3278347 by navneet0693: Junk plugin data can crash activity entity related view
Certain scenarios like:

1. configuration overrides
2. undefensive code
3. uninstalled modules not cleaning it's related data at the time of uninstalling

can lead to orphaned and/or junk data in entities in database. For example, a module declaring a plugin instance and attaching that instance with entities, for example, activities, was uninstalled; but during uninstall it didn't cleaned it's related data from activities. This will lead to PluginNotFoundException being thrown on the pages where activities are being processed. For example, home page, stream pages, group streams, notifications etc.

We wanted to make sure that before creating an instance of a plugin whose the plugin manager which extends 'DefaultPluginManager' class, we check that the plugin definition exists and it is successful in retrieving it without any errors.

The hasDefinition function is defined as follows:

public function hasDefinition($plugin_id) {
    return (bool) $this->getDefinition($plugin_id, FALSE);
}

which means that it is marking $exception_on_invalid parameter to be false, which simply says "If TRUE, an invalid plugin ID will throw an exception."

now hasDefinition simply returns false in this logic if the plugin definition couldn't be found. This saves us from hard errors and we can safely say that call to "createInstance" will return a plugin instance.