Added conditions to check for existing consent of users.
See: https://github.com/goalgorilla/open_social/pull/2913/files#r863616985

As said, as this feature was release already, and we are improving the hook_post_update_name later. There is possibility that users might have already given the consent. So, we want to preserve their preferences.

Added a new query to check for existing ids in the table. If there are ids existing we take a difference of ids and process them.