Add drupal/ginvite so we are using contrib instead of the custom module
We are using drupal/ginvite 1.1 as this closely matches the state of the fork. This should ensure we are backwards compatible.
After this is merged, we can decide to upgrade drupal/ginvite as well, so we are not just removing maintenance burden,
but also use features from the ecosystem and contribute back if we notice any issues that we encounter while upgrading.
But to make this step as small as possible we first move in to 1.1.