Add entity mapping to PHPStan
This uses a feature in PHPStan Drupal which makes it better aware of
what classes to load for the entity and storage of different entity
types. The list includes all entities found within the Open Social
distribution. In addition to this it includes classes from core or other
modules that we overwrite with our own class.

We may also add class mappings for contrib modules that we use that
don't have their own mapping.