Fix `always` issues in PHPStan
This tries to fix the PHPStan issues that contained `always` in the
baseline. In a few cases these indicated bugs in logic that have been
fixed. In most cases it cleans up the code and removes redundant checks
that the types tell us are not possible.

A common scenario is checking the return type of `getStorage` which
doesn't return `NULL` but would throw an exception instead.