Removed push event from the workflows.
The push event triggerred the workflows even after the merge. Our development workflow doesn't allows to push to main, only maintainers can do it. The workflow is to create a PR and then merge to main branch. The checks and tests already runs on PR, so there is no need of them to run again after merge to main.

Also, we remove manual work flow todo as Github actions is free for public repositories which means that we don't need to care about cost in terms of $$$. Thanks to Github for this <3. They really do love Open Source. See details at: https://github.com/features/actions#pricing-details. It was also confirmed by a response to ticket created by @bramtenhove to Github.