Ensure mergeable ignores translation updates
The translation updates are created by an automated system so we don't want mergeable to be triggered by them. A more stable thing to match for would be the branch name but I can't find in the documentation how to match that, so for now we just special case the title.