Issue #3308363 by rolki: Fix performance for custom content list block
The loading of results in the "Custom content list" block is very slow, especially on
some projects where the selection is more than 2000 items, after which the objects of
all these entities are loaded and a check is made to see if the user can see them,
this check is very slow.
The solution was to check access to the entity at the query level and at the same time
limit the number of objects, which seriously speeds up data processing.