Export screenshots for Behat test failures
We already have configuration that would create a screenshot if a test
failed. The only issue was that when we moved off of Travis this data
was no longer made available.

This commit ensures the GitHub actions upload the screenshots and logs
as job artifacts so that they can be viewed by developers.