Move Behat into a single workflow
We use the matrix function to run our different tests. Additionally we
switch from our Travis `stability` split up to just running all
individual features in separate jobs.

We take some inspiration from
https://tomasvotruba.com/blog/2020/11/16/how-to-make-dynamic-matrix-in-github-actions/
to discover all the feature files and then create a job for each of
them. This creates more noise in the list of tasks for a PR but it also
makes it a lot easier to figure out which task failed.