Make use of GitHub's Docker support
We added a `install-docker.sh` script a long time ago. However, GitHub
actually supports Docker so we can avoid a lot of the set-up we do
ourselves.

This should also ensure that GitHub caches any images we use which saves
on download time. It also shortens our workflows and makes some steps a
bit easier because we don't have to go into docker.