Create test job per feature folder rather than per feature
Creating a test job per feature causes ~200 jobs to be created. This
leads to GitHub spending more time on spinning up jobs than we're saving
with our parallelisation.

Moving to a job per feature folder reduces the number of jobs to about
20, increasing the speed in which they finish.

Sinc we removed `--fail-fast` in an earlier commit we still get output
for all tests in the first go, rather than having to work through one
failing test at a time.