Add web container to run tests against
We can not send web requests to the container that we're controlling in
our GitHub workflow. We can also not access any containers from our host
system. The trick we end up with is to add a web container that will
actually serve as our webserver. We keep running all our commands in a
container too so that we're connected to the network. This is similar to
the `behat` container that we had in our previous setups.

Previously we tried one-or-the-other but it turns out we need both.