Make flood-edit test deterministic
Since we've added a `Configuration` item in the site manager menu, if
the test starts before the admin menu has collapsed then Behat may press
the wrong "Configuration" item. Since we don't want to test the menu
structure of Drupal, it's safe to go to the path directly to ensure this
functionality works.