Rewrite book create test
We rewrite the book create test so the scenario looks more like
acceptance criteria that we can discuss with stakeholders.

The nitty gritty of efficiently mapping to fields is abstracted away in
the context. The context should try to be as implementation agnostic as
possible, though there are some things we need to be specific about
(like multiple title fields and the inner workings of CKEditor).