Remove click-avatar profile test
This is functionality that hasn't regressed in the life of Open Social
and the test hasn't been run in a long time since it was missing the
`@stability` tag. Now that we run tests without tags we found the test
was broken.

Removing the test is the easiest option especially since it apparently
offered no value.