Transform Book in Flexible group to readable test
We apply the learnings from our previous commits to move the book in
flexible group test into a format that can easily be used to communicate
with stakeholders.

While doing this we fix the project/module definitions in some affected
files to resolve a bug uncovered by the test.

This test wasn't running in the previous `@stability-*` set-up because
it did not have the pre-requisite tag. However, the test is now running
since all test files are executed.

We update the BookContext as needed and move some shared functionalities
into GroupTrait and NodeTrait so that we can re-use helper functions
without having to fetch additional contexts. Those traits are not
supposed to contain any steps though since they would be duplicated.