Improve test stability and simplicity
We improve a few steps so we can wait for actual UI updates rather than
for specific implementation events which we sometimes miss (e.g. look at
the effects of AJAX rather than try to catch the AJAX request itself).
This should make the tests more stable.

Additionally we change how we create some of the entities at the start
of a test. We now validate that the provided information is actually
valid for the entity. This prevents later assertions from saying text
isn't there because a field name was mistyped.