Make chrome the default browser
This way at least we know that JavaScript will work even for tests that
don't have the `@javascript` tag, because it seems like some tests need
JavaScript but don't declare their dependency.

We could add the `@javascript` tag to those tests but that currently
causes an issue with the DrupalExtension Context in some cases that's
also difficult to debug.