Add EntityTrait for field creation vlaidation
This allows us to validate that we're using the right fields when we're
creating an entity. While the previously added `validate` call ensures
all required fields are set with the right values, this added function
checks we're not trying to set fields that don't exists since this may
cause tests to fail for unexpected reasons (e.g. due to a typo in a
field name).