Revert "Added translation using Weblate (Bulgarian)"
This reverts commit 958d09ba77f48595cb43de64e6e94d51d0aa214b.

These languages were automatically added by Weblate which to sync the
languages from the mobile app repositories which were still in the Open
Social project but use a different (flat) folder structure for
translations.

These languages should not exist in the `/` of translations because
they're split out to the individual `/gp*` subfolders mapping to
greenpeace specific Weblate projects.