array_search returns the key, however in the case of indexed array this is the index the search was found and we unset by key not index.
You can find more information in #3316206, where you will see that the array was [0 => [], 2 => [], 3 => []] and array_search returned 2
which means, the second element within the array which starts at 0. So that would be the element with key 3.
Unsetting meant that actually the first element with key 2 was unset. Resulting in the wrong VBO action being removed from the list.

I have also assigned a new variable for the array_column. So that it doesn't call array_column for each element it searches.