Ignore an existing bug found by our tests
Even though we should not be extending the list of ignored bugs for new
features we write, we unfortunately can not fix every bug we discover by
improving existing tests. So for now we report the bug and ignore it to
ensure our test can still provide improved coverage for other parts.