Require an author for scaffolded groups in tests
We were allowing the author to be omitted in tests which makes writing
tests easier. This prevents users being created as parts of the test
steps even though they were not technically needed for the tests.

The downside of this is that groups were associated with the anonymous
user which could cause issues themselves for tests involving anonymous
viewers.

To solve both of these issues we make the author required and need tests
to be explicit about who owns the groups. We do provide two new way of
scaffolding groups which allows communicating that the owner themselves
doesn't matter but should just be anonymous, or to indicate that the
owner does matter but can be the current user. In other cases the owner
should be specified specifically and is apparently important for the
test.