Implement access check for group content editing
This commit implements the current rules for content editing within
groups which is mostly gleaned from our current documentation on
https://communitytalks.getopensocial.com/manual-features-and-site-settings/groups

We do not test the actual editing of content itself because that's part
of the feature for which we create the content. We also do not test that
the group of content can be changed. That is tested in a separate file.