Fix PHPStan issues
In a previous commit we've started using the `$entity` for an entity
view hook but PHPStan doesn't know this is a specific entity class
rather than just any entity class.

We add an assert because we only expect that entity class to be provided
for a properly named hook, as per Drupal's hook system.

We also slightly rewrite flexible_group code to get the proper type for
our account id which is needed until
https://www.drupal.org/project/drupal/issues/3324733 is fixed.