Run on all branches except a few rather than only main
The risk of running only on `main` is that we break a release branch in
the form of `x.y.z` unknowingly. GitHub doesn't allow us to use regex
for branch names.

So instead we run on all branches but we include patterns that we've
seen being used for development branches. This ensures we don't run on
correctly named development branches while running on anything that may
be a release branch. There's a chance we run on a misnamed development
branch and perform duplicate work of the pull request, but that's less
of a risk than missing a release branch.