Allow test-folders to be nested for test discovery
We want to be able to split up our jobs in a more fine-grained manner to
have better control over how long the test suite runs (e.g. we currently
have a single feature file that takes 35 minutes we'd lke to break up).

This adjusts the feature discovery to find folders and subfolders that
have at least one .feature file in them. This allows us to create nested
groups of feature files to separate them out to different jobs while
keeping the context when working on the tests.