Move flexible groups tests into groups folder
Since parallelism is no longer limited to top level folders but now also
properly handles sub-folders, we move the flexible group tests into the
groups folder. This makes them easier to be discovered and documents
that it's a sub-part of this specific feature.

I use this opportunity to rename `group` to `groups` ot differentiate
between our feature (`groups`) which we're testing and the module
(`group`) which has its own tests and we trust to work correctly.