Rewrite the group edit group type test
This test was failing randomly in some of our automated testing with an
unclear reason.

The rewrite aims to have the same coverage but uses some of our new
testing primitives introduced in the flexible groups rewrite to make the
test more compact and readable. We also add a new assertion to properly
expand the settings tab that contains the group type settings field
which should ensure the test no longer fails randomly.