Run translations workflow on stable PHP
We were setting PHP to 7.4 because we did not yet have 8.x support.
However we now support 8.0 which is the default in the GitHub runner.

GitHub has stopped shipping PHP 7.4 in their runners because that
version is EOL as of 28th of November of this year.

We remove setting a specific version because I'm confident we'll be able
to keep everything up to date that we support the next PHP versions
(e.g. 8.1, or 8.2) if and when GitHub updates their runners.