Fix tugboat failing for updated dependencies
Tugboat prefers updating over recreating (since it's faster). However, our
composer command did not allow updates of dependencies which meant
that the Tugboat preview update fails if we update our dependencies.

This can cause Tugboat to fail for dependabot but also when we finally
merge our dependency update to main, which then causes Tugboat to
fail on all PRs (which use the main preview as a base to save disk space).

We change our composer require to allow updating dependencies so that
this scenario no longer occurs.