Revert change from 215c7db
In PR 2962 we introduce a dependency towards graphql_oauth, this means it will enable simple_oauth while we still use a different module for the native app, namely oauth2_server.
This causes the native app to break, also the simple_oauth will be enabled without any configuration. As some point we want the simple_oauth to take over, but in a smooth way.

Also at first it will start calling a function in getExtension while there is nothing enabled yet and also breaking that.