Use a dev tools version compatible with multiple major versions
It's difficult to figure out which dev tools version belongs to which
major. So instead we're changing around the open_social_dev package
itself to be compatible with the two latest major versions at the same
time and we ensure a CI inside of that repository enforces that.

That makes our upgrade path test a lot simpler.

We allow patch level upgrades since this is a package we control
ourselves so we can properly follow SemVer.