Fix issue 3302484 for address module
commerceguys/addressing introduces some type changes in 1.4.0 with
https://github.com/commerceguys/addressing/commit/c81ded457fa0cc7d2c9f5b05294909b20da1b2dc

However, that's actually a breaking change for drupal/address which
means that anything below 1.11 no longer works with that version.
(https://www.drupal.org/project/address/issues/3302484)

We can not update the drupal/address module without creating a new minor
version which we no longer want to do. However, as part of a bugfix we
can prevent installation of the broken version so that installing ^10.3
results in a working Open Social installation.