Rewrite feature to clarify the goal of the feature itself
I've rmeoved `Benefit`/`Role`/`Goal` from the test for a few reasons.

Overall at Open Social we're not great at filling out these fields so
the information is limited. The design team itself works with Jobs to be
Done, which would be more useful metadata. We could add that if it
exists for a feature.

There's also a pitfall in `Role:` specifically because Behat (or
Mink/Drupal Mink) will try to log in as a user with that role, so this
adds set-up to your test that is less explicit than our other steps.