Prefer 'given logged in as' over manual log-in
Using the shorthand has a few advantages:

1. The only real thing we care about is that the user is a site manager.
The functionality should change even if we get rid of `benflorez` as a
user (or change his permissions). Logging in as a site manager rather
than a specific user makes this clear.

2. The "logged in as" can keep working even if we change how our login
works. For example we may switch to passwordless login, or require
authentication through SSO. Or we may change how our login form works.

3. It's faster and more readable. We're skipping a few test steps that
take up space on our screen and we avoid a few web requests making the
test faster.