Additional feedback
The following things take a bit longer to fix so they don't have a
`diff` and they're also code formatting choices that we can debate.

I've added a blank line between `Given`, `When`, `Then` blocks to
differentiate between the three phases of a [test: `Arrange`, `Act`,
`Assert`.](https://automationpanda.com/2020/07/07/arrange-act-assert-a-pattern-for-writing-good-tests/)

I'd also prefer to remove the nesting of `And` statements. It's a
nesting that's not present in all of our other tests so consistency is
my first argument. The second argument is that the steps are usually of
equal importance and having them on the same line makes readbility
easier (though that's subjective and debatable).

I think the third scenario in the written feature also needs some work.
I'm of the strongly held believe  that all scenario's should be
independent. We currently load a clean database for every `.feature`
file. The reasson we don't do this for a scenario is mostly a current
technical limitation that we may solve in the future.

Ensuring scenario's are independent makes them easier to debug because
it limits the scope of what I need to look at and keep in mind. It also
means that all arrangements on top of a blank installation are made
explicit.

Independent scenario's also ensure we can easily move them around as
needed.

The purist in me would like to split the second scenario into 2, so that
we only have a single "Arrange, Act, Assert" rather than the current
"Arrange, Act, Assert, Act, Assert". I think it's fine ot have a test
for the form that then checks the right config was updated and a
separate test that arranges certain config and then checks the user
profile editing.

While writing a test in that way may violate the "black box" principle
somewhat it provides more clarity over what is broken, when things are
broken (i.e. did the configurtion not save correctly, or is the form not
using the right configuration).

Finally, we should avoid using `id`s for our tests (i.e. "I check the
box <id>"). id's are usually generated by Drupal which means they may
change without the functionality changing, making our tests less
resilient than they can be. More importantly, our platform users have no
knowledge of IDs and ensuring we test using labels also ensures we have
an accessible product. If a test can not be written using a label, then
that's a good indicator that we have an accessibility issue in our
product that should be fixed :)