Internal: Remove redundant highlight dependency
We are not using this library in code base anywhere. In the pages of history, this was somewhere used in Open Social style guide which we do not support anymore.

This was first introduced in https://github.com/goalgorilla/open_social/pull/1358. It was added as dependency of KSS in our socialblue theme which can be seen here: https://github.com/goalgorilla/open_social/pull/1358/files#diff-f688afcb60f350ef698bb13a058e38938e27379d4fc5ec1152b4fdecf1ae2de7R3451

This library helped us in having highlighted snippets in our editors like code block:

```
<pre><code class="language-markup" data-language="html">{{ section.markup|escape('html') }}</code></pre>
```

We are removing it as we don't require it any more.